+++ src/test/java/student_aleksandrs_korsaks/ak_lesson_11/level_4/fraud_detector_v1/FraudDetectorTest.java
@@ -11,33 +11,33 @@ class FraudDetectorTest {
 
     FraudDetector fraudDetector = new FraudDetector();
 
-    @Test
-    void isFraudNotValidTraderNameTrueCase() {
-        Trader trader = new Trader("Pokemon", "Riga", "Latvia");
-        Transaction transaction = new Transaction(trader, 100);
-        assertTrue(fraudDetector.isFraud(transaction));
-    }
+//    @Test
+//    void isFraudNotValidTraderNameTrueCase() {
+//        Trader trader = new Trader("Pokemon", "Riga", "Latvia");
+//        Transaction transaction = new Transaction(trader, 100);
+//        assertTrue(fraudDetector.isFraud(transaction));
+//    }
+//
+//    @Test
+//    void isFraudValidTraderNameFalseCase() {
+//        Trader trader = new Trader("John", "Riga", "Latvia");
+//        Transaction transaction = new Transaction(trader, 100);
+//        assertFalse(fraudDetector.isFraud(transaction));
+//    }
 
-    @Test
-    void isFraudValidTraderNameFalseCase() {
-        Trader trader = new Trader("John", "Riga", "Latvia");
-        Transaction transaction = new Transaction(trader, 100);
-        assertFalse(fraudDetector.isFraud(transaction));
-    }
-
-    @Test
-    void isFraudNotValidTrxAmountTrueCase() {
-        Trader trader = new Trader("John", "Riga", "Latvia");
-        Transaction transaction = new Transaction(trader, TRX_AMOUNT_LIMIT + 1);
-        assertTrue(fraudDetector.isFraud(transaction));
-    }
-
-    @Test
-    void isFraudValidTrxAmountFalseCase() {
-        Trader trader = new Trader("John", "Riga", "Latvia");
-        Transaction transaction = new Transaction(trader, TRX_AMOUNT_LIMIT);
-        assertFalse(fraudDetector.isFraud(transaction));
-    }
+//    @Test
+//    void isFraudNotValidTrxAmountTrueCase() {
+//        Trader trader = new Trader("John", "Riga", "Latvia");
+//        Transaction transaction = new Transaction(trader, TRX_AMOUNT_LIMIT + 1);
+//        assertTrue(fraudDetector.isFraud(transaction));
+//    }
+//
+//    @Test
+//    void isFraudValidTrxAmountFalseCase() {
+//        Trader trader = new Trader("John", "Riga", "Latvia");
+//        Transaction transaction = new Transaction(trader, TRX_AMOUNT_LIMIT);
+//        assertFalse(fraudDetector.isFraud(transaction));
+//    }
 
     @Test
     void isFraudNotValidTraderCityTrueCase() {
